rewrite menu code
The menu handling functionality is moved to a class `MatrixMenu` in
Matrix/menu.py.  Printing the menus and reading the user's selection are
moved to separate methods The menus functions are cleaned up, some
redundancy is removed, and more error handling is introduced.  The
protocol configuration file path is read only once and then saved for
the session.  However, it can be changed by an option in the main menu.