load protocol config file once by the menu
The protocol configuration is now read in the Menu part of the
application after the user has specified the corresponding path.  The
parsed dictionary is then given to the other components.  However, the
path is still given to the Execution module since it is required by the
fabfile.  This adds also basic error handling, e.g., if the given
filename is not accessable.