fix AWS price bids with multiple regions
When iterating through the regions, the variable `price_bids` containing
the highest acceptable price of spot instances was changed to the a
string representation of the minimum of its current value and the
winning bid price of the current region.

This failed in case of more than one region: The comparison between
string and float failed in the next iteration.  Also, sometimes a lower
bid price would be used in the next region.

Additionally, the return type of `check_latest_price` is changed to
float.